+++ src/models/tabauto/keras_model.py
@@ -1,9 +1,26 @@
+# Copyright 2024 IBM Corp.
+# 
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+# 
+#     http://www.apache.org/licenses/LICENSE-2.0
+# 
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 import os
 import matplotlib.pyplot as plt
 from tensorflow.keras.models import Sequential
 from tensorflow.keras.models import load_model
 from tensorflow.keras.layers import Dense, Flatten, Conv1D
-from tensorflow.keras import optimizers, losses
+from tensorflow.keras import losses
+
+import tensorflow.keras.optimizers.legacy
+
 from tensorflow.keras.callbacks import (
     LearningRateScheduler,
     EarlyStopping,
@@ -75,9 +92,10 @@ class KerasModel(BaseModel):
                 outputs=output_node,
                 directory=tmp_path,
                 max_trials=self.max_trials,
-                objective="val_loss",
+                objective=["val_loss"],
                 tuner=self.tuner,
                 seed=self.random_state,
+                optimizer=tensorflow.keras.optimizers.legacy.Adam(),
             )
 
         else:
@@ -96,9 +114,10 @@ class KerasModel(BaseModel):
                 outputs=output_node,
                 directory=tmp_path,
                 max_trials=self.max_trials,
-                objective="accuracy",
+                objective=["accuracy"],
                 tuner=self.tuner,
                 seed=self.random_state,
+                optimizer=tensorflow.keras.optimizers.legacy.Adam(),
             )
 
         self.model = model
@@ -223,7 +242,7 @@ class KerasModel(BaseModel):
                     )
 
         # choose optimizer
-        opt = optimizers.Adam()
+        opt = tensorflow.keras.optimizers.legacy.Adam()
 
         # choose loss function
         if self.dataset_type == REGRESSION: