+++ src/utils/utils.py
@@ -1,8 +1,17 @@
-# --------------------------------------------------------------------------
-# Licensed Materials - Property of IBM
+# Copyright 2024 IBM Corp.
 #
-# (C) Copyright IBM Corp. 2019, 2020
-# --------------------------------------------------------------------------
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 
 import cProfile
 import glob
@@ -161,7 +170,7 @@ def setup_logger(experiment_folder):
 
     lg_file["handlers"]["file"]["filename"] = str(
         experiment_folder
-        / f"AutoOmicLog_{str(int(datetime.timestamp(datetime.utcnow())))}.log"
+        / f"OmiXaiLog_{str(int(datetime.timestamp(datetime.utcnow())))}.log"
     )
     logging.config.dictConfig(lg_file)
     omicLogger = logging.getLogger("OmicLogger")